# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing sync operations",
        "title": "Beyond Sync API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/sync": {
            "get": {
                "description": "Retrieves all sync records from Redis",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sync"
                ],
                "summary": "Get all sync records",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handlers.SyncResponse"
                            }
                        }
                    }
                }
            }
        },
        "/sync/trigger": {
            "post": {
                "description": "Triggers a new synchronization operation",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sync"
                ],
                "summary": "Trigger new sync operation",
                "responses": {
                    "201": {
                        "description": "Sync trigger is requested",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/sync/{timestamp}": {
            "get": {
                "description": "Retrieves a specific sync record by its timestamp",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sync"
                ],
                "summary": "Get sync record by timestamp",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Timestamp of the sync record",
                        "name": "timestamp",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.SyncResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handlers.SyncResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/redis.RedisValue"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
        "redis.RedisValue": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "sha": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        }
    }
}